requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.12.1+cu113 # --index-url https://download.pytorch.org/whl/cu113
torchvision==0.13.1+cu113 # --index-url https://download.pytorch.org/whl/cu113
diffusers[torch]==0.11.1
accelerate==0.15.0
transformers==4.25.1
bitsandbytes==0.35.4
einops
omegaconf
ftfy
tensorboard
modelcards
imageio
triton
click
opencv-python
imageio[ffmpeg]
six
scikit-image
datasets
tqdm
regex
pyyaml
numpy
axial_positional_embedding
timm
einops_exts
ninja
matplotlib
torchdiffeq
python-etcd
Jinja2
joblib
